+++ kitty/rc/launch.py
@@ -2,7 +2,7 @@
 # License: GPLv3 Copyright: 2020, Kovid Goyal <kovid at kovidgoyal.net>
 
 
-from typing import TYPE_CHECKING, Optional
+from typing import TYPE_CHECKING, Dict, Optional
 
 from kitty.cli_stub import LaunchCLIOptions
 from kitty.launch import launch as do_launch
@@ -21,7 +21,7 @@ class Launch(RemoteCommand):
     args+/list.str: The command line to run in the new window, as a list, use an empty list to run the default shell
     match/str: The tab to open the new window in
     window_title/str: Title for the new window
-    cwd/str: Working directory for the new window
+    cwd/str=copy_local_cwd: Working directory for the new window
     env/list.str: List of environment variables of the form NAME=VALUE
     var/list.str: List of user variables of the form NAME=VALUE
     tab_title/str: Title for the new tab
@@ -29,7 +29,7 @@ class Launch(RemoteCommand):
     keep_focus/bool: Boolean indicating whether the current window should retain focus or not
     copy_colors/bool: Boolean indicating whether to copy the colors from the current window
     copy_cmdline/bool: Boolean indicating whether to copy the cmdline from the current window
-    copy_env/bool: Boolean indicating whether to copy the environ from the current window
+    copy_env/list.str=copy_local_env: List of strings representing the local env vars
     hold/bool: Boolean indicating whether to keep window open after cmd exits
     location/choices.first.after.before.neighbor.last.vsplit.hsplit.split.default: Where in the tab to open the new window
     allow_remote_control/bool: Boolean indicating whether to allow remote control from the new window
@@ -83,17 +83,30 @@ instead of the active tab
         default_opts = parse_launch_args()[0]
         opts = LaunchCLIOptions()
         for key, default_value in default_opts.__dict__.items():
+            if key == 'copy_env':
+                continue
             val = payload_get(key)
             if val is None:
                 val = default_value
             setattr(opts, key, val)
+        ceval = payload_get('copy_env')
+        opts.copy_env = False
+        base_env: Optional[Dict[str, str]] = None
+        if ceval:
+            if isinstance(ceval, list):
+                base_env = {}
+                for x in ceval:
+                    k, v = x.partition('=')[::2]
+                    base_env[k] = v
+            elif isinstance(ceval, bool):
+                opts.copy_env = ceval
         target_tab = None
         tabs = self.tabs_for_match_payload(boss, window, payload_get)
         if tabs and tabs[0]:
             target_tab = tabs[0]
         elif payload_get('type') not in ('background', 'os-window', 'tab', 'window'):
             return None
-        w = do_launch(boss, opts, payload_get('args') or [], target_tab=target_tab, rc_from_window=window)
+        w = do_launch(boss, opts, payload_get('args') or [], target_tab=target_tab, rc_from_window=window, base_env=base_env)
         return None if payload_get('no_response') else str(getattr(w, 'id', 0))